+++ pennylane/beta/devices/default_tensor.py
@@ -153,7 +153,7 @@ class DefaultTensor(Device):
 
     _zero_state = np.array([1.0, 0.0], dtype=C_DTYPE)
 
-    def __init__(self, wires, shots=1000, representation="exact", contraction_method="auto"):
+    def __init__(self, wires, shots=None, representation="exact", contraction_method="auto"):
         super().__init__(wires, shots)
         if representation not in ["exact", "mps"]:
             raise ValueError("Invalid representation. Must be one of 'exact' or 'mps'.")
@@ -164,7 +164,9 @@ class DefaultTensor(Device):
         self._rep = representation
         self._contraction_method = contraction_method
         self.reset()
-        self.analytic = False
+
+        if shots is not None:
+            raise ValueError("Device can only perform analytic computations.")
 
     @classmethod
     def capabilities(cls):