+++ tests/test_optimize.py
@@ -250,7 +250,7 @@ class TestOptimizer:
         possible_generators = [qml.RX, qml.RY, qml.RZ]
         bunch.rotoselect_opt.possible_generators = possible_generators
 
-        dev = qml.device("default.qubit", analytic=True, wires=2)
+        dev = qml.device("default.qubit", shots=None, wires=2)
 
         def ansatz(params, generators):
             generators[0](params[0], wires=0)
@@ -629,7 +629,7 @@ class TestOptimizer:
         possible_generators = [qml.RX, qml.RY, qml.RZ]
         bunch.rotoselect_opt.possible_generators = possible_generators
 
-        dev = qml.device("default.qubit", analytic=True, wires=2)
+        dev = qml.device("default.qubit", shots=None, wires=2)
 
         def ansatz(params, generators):
             generators[0](params[0], wires=0)
@@ -670,7 +670,7 @@ class TestOptimizer:
         possible_generators = [qml.RX, qml.RY, qml.RZ]
         bunch.rotoselect_opt.possible_generators = possible_generators
 
-        dev = qml.device("default.qubit", analytic=True, wires=2)
+        dev = qml.device("default.qubit", shots=None, wires=2)
 
         def ansatz(params, generators):
             generators[0](params[0], wires=0)